Move App.js to the src folder
