+++ src/tests/App.test.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from '../../redux/configureStore';
-import App from '../../components/App';
+import store from '../redux/configureStore';
+import App from '../App';
 
 test('Check if the App is rendering correctly', () => {
   render(
